requirements.txt
streamlit>=1.31
pandas>=2.0
openpyxl>=3.1
plotly>=5.18
requests>=2.31
bcrypt>=4.1
python-dotenv>=1.0
